Adding insertion order to all object
As part of the refactoring of even ordering that I need to do to
make sure that multi-call MIDI events (such as NRPN calls) are in the
correct order, I've added the ability for all events to keep
track the order they've been placed on the list. This will disambiguate
ordering when time and ordinal are the same.