Throw away the MIDIEvent class
Add a tick attribute to GenericEvent.

Change the name of the Note class to NoteOn, and create a new class NoteOff.

Change addNoteByNumber() to append both a NoteOn and corresponding NoteOff
to the MIDITrack.eventList.

Simplify the big mess that was processEventList().

All the knowledge of the serialization form of each kind
of event is removed from writeEventsToStream() and put where it
belongs, in each derived class of GenericEvent.

The big mess which was in writeEventsToStream() is greatly simplified.